refactor(base_type_copilot): split into mod+transcript (#1266)
base_type_copilot.rs 578 -> mod.rs 368 + transcript.rs 213.
Transcript parsing helpers (extract_response_from_transcript,
is_copilot_footer_line, is_transcript_noise_line, strip_ansi)
moved to transcript.rs and re-exported.
